+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ### 🐞 Fixes
 
+- [#629](https://github.com/estruyf/vscode-front-matter/issues/629): Fix array indent to the new property
 - [#660](https://github.com/estruyf/vscode-front-matter/issues/660): Allow only to select unique content relationship values
 - [#661](https://github.com/estruyf/vscode-front-matter/issues/661): Fixing the dropdowns when used at the bottom of a collapsible group
 - [#664](https://github.com/estruyf/vscode-front-matter/issues/664): Fix for parsing draft status in Hexo and Jekyll